+++ src/main/java/com/sparta/todo/controller/LikeController.java
@@ -1,16 +1,10 @@
 package com.sparta.todo.controller;
 
-import com.sparta.todo.dto.SuccessMessageDto;
-import com.sparta.todo.dto.request.PostRequestDto;
-import com.sparta.todo.dto.request.ToDoRequestDto;
-import com.sparta.todo.dto.response.LikeResponseDto;
-import com.sparta.todo.dto.response.PostResponseDto;
-import com.sparta.todo.entity.User;
+import com.sparta.todo.dto.requestDto.PostRequestDto;
+import com.sparta.todo.dto.responseDto.PostResponseDto;
 import com.sparta.todo.security.UserDetailsImpl;
 import com.sparta.todo.service.LikeService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-//import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,10 +17,6 @@ public class LikeController {
     public PostResponseDto LikePost(@PathVariable Long postId, @RequestBody PostRequestDto postRequestDto, @AuthenticationPrincipal UserDetailsImpl userDetails){
         return likeService.likePost(postId,postRequestDto, userDetails.getUser());
     }
-//
-//    @PutMapping("/like/{postId")
-//    public ResponseEntity<LikePostResponseDto> LikePost(@PathVariable Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails){
-//        return likeService.likePost(postId, userDetails.getUser());
-//    }
+
 
 }
