--- src/main/java/com/sparta/todo/dto/request/LikeRequestDto.java
@@ -1,9 +1,0 @@
-package com.sparta.todo.dto.request;
-
-
-import lombok.Getter;
-
-@Getter
-public class LikeRequestDto {
-
-}
