+++ src/main/java/com/sparta/todo/dto/requestDto/LoginRequestDto.java
@@ -11,6 +11,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class LoginRequestDto {
     private String email;
-
     private String password;
 }
