+++ src/main/java/com/sparta/todo/dto/requestDto/Request.java
@@ -1,14 +1,13 @@
-package com.sparta.todo.dto.request;
+package com.sparta.todo.dto.requestDto;
 
 import com.sparta.todo.entity.Category;
 import lombok.Getter;
 
-
 @Getter
-public class ToDoRequestDto {
-    private Long toDoId;
+public class Request {
+    private String date;
     private Category category;
     private String content;
     private Boolean done;
-    private Long postId;
+
 }
