+++ src/main/java/com/sparta/todo/dto/requestDto/ToDoRequestDto.java
@@ -1,13 +1,14 @@
-package com.sparta.todo.dto.request;
+package com.sparta.todo.dto.requestDto;
 
 import com.sparta.todo.entity.Category;
 import lombok.Getter;
 
+
 @Getter
-public class Request {
-    private String date;
+public class ToDoRequestDto {
+    private Long toDoId;
     private Category category;
     private String content;
     private Boolean done;
-
+    private Long postId;
 }
