+++ src/main/java/com/sparta/todo/repository/ToDoRepository.java
@@ -1,6 +1,5 @@
 package com.sparta.todo.repository;
 
-import com.sparta.todo.entity.Category;
 import com.sparta.todo.entity.Post;
 import com.sparta.todo.entity.ToDo;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,6 +11,5 @@ import java.util.List;
 public interface ToDoRepository extends JpaRepository<ToDo, Long> {
 
     List<ToDo> findAllByPostOrderById(Post post);
-    List<ToDo> findAllByCategory(Category category);
 
 }
