+++ src/main/java/com/sparta/todo/repository/UserRepository.java
@@ -10,7 +10,6 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
-
     Optional<User> findByNickname(String nickname);
 
 }
